[location-component] Location component animation (#11)
* Introduce animation system

* Fixes after rebase

* PR fixes

+++ plugin-locationcomponent/src/main/java/com/mapbox/maps/plugin/locationcomponent/LocationLayerRenderer.kt
@@ -23,7 +23,7 @@ internal interface LocationLayerRenderer {
 
   fun setLatLng(latLng: Point)
 
-  fun setBearing(bearing: Float)
+  fun setBearing(bearing: Double)
 
   fun setAccuracyRadius(accuracy: Float)
